requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
requests==2.32.3
streamlit==1.40.2
Pillow==10.4.0
openai==1.57.0
